4ta entrega + intento de persistencia

+++ src/aterrizar/Filtro.java
@@ -4,5 +4,5 @@ import java.util.ArrayList;
 
 public interface Filtro {
 	ArrayList<Asiento> filtrar(ArrayList<Asiento> asientos, String clase, String ubicacion, 
-							Float precioMin, Float precioMax,Boolean conReservados);
+							Double precioMin, Double precioMax,Boolean conReservados);
 }
